pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Kobina Brandon <maadabrandon@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
polygon-api-client = "^1.13.4"
python-dotenv = "^1.0.1"
websocket = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
